cpufreq: interactive: Reduce DEFAULT_GO_MAXSPEED_LOAD
Google increased it to 95 in order to save battery, but that only
worked because a bug causes the code for the scaler to scale down
to also scale up. With that bug fixed, the phone will have trouble
scaling up with go_maxspeed_load set to such a high value.